Shaders/materials fixed, assets added, bugs discovered.
Important notes: I removed the renderer element for the ground boxes. they are useless cuz the boxes are transparent anyway and were taking gpu space for nothing. Fixed map planes, now it's one object modeled from blender. All planes in prefabs are deleted (kept colliders ofc). With that, ground material fixed, it looks cool now. IMPORTANT NOTE HERE: you will need to change the objective thing in the script if you called the cube. I deleted the cube and added a diamond with a special material on it. Call that instead. Found a bug that may be responsible for the horrible lag, talk to u in discord to explain cuz idk how to here. Yallah sir t7awa.